fix(lampod): move the runtime call down to the peer stack
This fixes the issue mensioned in [1] where we was constraint
to use tokio to all `lampod::call` but this is necessary only
to run the `peer` operation.

So we move the async call down to the stack and we remove a few
of the async contraint in the usage of the API.

Like magic we the error that we can block a task
when there are async operation that are going on is gone.

[1] https://github.com/dart-lightning/lampo.rs/issues/29

Link https://github.com/dart-lightning/lampo.rs/issues/29
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>