feat(lnprototest): add an initial implementation of lnptototest runner
lnptorotest is an opinionated Lightning BOLT Protocol Test Framework developed
by Rusty Russell (and I’m the official maintainer of it) that point to provide a
test framework for a lightning network implementation such as core lightning.

So this commit will introduce a basic implementation of the runner
to support lampo and make a simple init interop test.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>