feat(c-ffi): introduce the basic lampod API through C FFI
This introduce the basic API to add the possibility to embedd
the lampod code inside any language through the C FFI.

This is an experimental version of the package because I do not
have any experience in writing these kind of C FFI bindings.

So all the work is based on [1] that make my life easy!

[1] https://bheisler.github.io/post/calling-rust-in-python

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>