feat(bitcoind): add support for bitcoind
This commit will add the possiblity to register the
bitcoin core backend (it is still incompleted) but
this should allow up to work with it.

Due the mismatch with nakamoto and all the bitcoin
ecosystem that run the bitcoin 0.30 thare are
a lot of hacks around the code that should be fixed
as soon as we get nakamoto updated.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>