feat: implement the event passing to update ldk status with onchain
The async programming (not the rust language) with the ldk is a kinda
hard because you need to handle different things at the same time.

So I am proposing an event programming API to allow to update ldk
with events and generated by the lampo handler from different part of
the application.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>