fix(bitcoind): Corrects block analysis for better accuracy
We identified a bug in our lampo-bitcoind implementation that
was causing the LDK (Lightning Development Kit) to force-close
channels. This bug stemmed from the code used to analyze transactions
within the blockchain during block analysis.

This commit addresses the issue and also introduces improvements
for handling transactions submitted from LDK to our backend.
Specifically, a watch_transaction that is broadcasted by us
(for instance, when opening a channel) will no longer be retained in
the external_tx vector.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>