cli: ask as an input the bip 39
Before this commit the mnemonic was passed by
command line option and it was a privacy leak
because the information was stored inside the
terminal history.

With this commit, we ask the seeds as input each
time that lampo is run with `--restore-wallet` option.

In the future, this option will be required only the
first time of initialization of the wallet.

Link: https://github.com/vincenzopalazzo/lampo.rs/issues/211
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>
Co-authored-by: Harshit <harshitverma933@gmail.com>