lnprototest: fix BAD MAC error
We was forcing the public key of CLN derived from the
private key that we was enforcing. However, every implementation
will derive the public key in in different way and we should
pass down a public key that is derived from the node itself.

In this case we should call just `lampo-cli getinfo` and use the
public key that is returned by the command.

Developed-by: @Psycho-Pirate
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>