Wait time added for log files
This commit adds wait time for log files to get generated. This avoids the `FileNotFoundError: [Errno 2] No such file or directory`. This is needed because script launches the daemon using a nohup ... & command, but immediately tries to read the log files before the daemon has a chance to create them.