core: migrate channels to support async
Initially, Lampo was synchronous and relied on a
custom, hand-written library to handle asynchronous
functionality such as RPC. Over time, it became
clear that Lampo should fully embrace asynchronous
programming to avoid making the codebase
increasingly difficult to maintain.

Currently, some parts of the code use channels to
allow users of the Lampo SDK to receive events
via the `handler.handler()` function, which sends
events through a channel.

At the time, `crossbeam_channel` was chosen for
this purpose. However, as we transition to asynchronous
code, we need to adopt a solution that supports
async operations.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>